fix!: (Barely tested) Make compatible with latest probe.py changes
In commit d4bae4d, as part of a larger group of changes[1], significant
pieces in the way probes are implemented and used has been reworked.
These changes accommodate that rework but are backwards incompatible
with earlier versions.

I have tested this on a single machine (auto-Z, non-virtual-endstop),
please test your printer with these changes before proceeding as normal
(if you break your printer, that's not on me).